[tests] Promote run_async_fn fixture to pytest conftest.py in preparation for further re-use.
